separate binding generator
